refactor(ssm): replace GDN files with PR 5765
Use the GDN package and unit-test files from NVIDIA/Megatron-LM PR #5765 as the split-file baseline before replaying dev-side GDN changes.
