+++ tests/unit_tests/ssm/test_gated_delta_net.py
@@ -1,26 +1,24 @@
 # Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 
+import copy
 import os
-from functools import partial
-from unittest import mock
 
 import pytest
 import torch
 import torch.nn.functional as F
 
 from megatron.core import parallel_state
-from megatron.core.models.common.embeddings.rope_utils import (
-    get_pos_emb_on_this_cp_rank as get_tensor_on_this_cp_rank,
-)
 from megatron.core.models.gpt.experimental_attention_variant_module_specs import (
     get_experimental_attention_variant_module_spec,
     get_transformer_block_with_experimental_attention_variant_spec,
 )
-from megatron.core.models.gpt.gpt_model import GPTModel
-from megatron.core.packed_seq_params import PackedSeqParams
 from megatron.core.process_groups_config import ProcessGroupCollection
 from megatron.core.ssm.gated_delta_net import (
+    HAVE_FLA,
     GatedDeltaNet,
+    torch_chunk_gated_delta_rule,
+)
+from megatron.core.ssm.gated_delta_net.common import (
     _build_head_perm_for_split_sections,
     _build_thd_cp_a2a_perm,
     tensor_a2a_cp2hp,
@@ -28,29 +26,12 @@ from megatron.core.ssm.gated_delta_net import (
 )
 from megatron.core.tensor_parallel.random import model_parallel_cuda_manual_seed
 from megatron.core.transformer import TransformerConfig
-from megatron.core.utils import unwrap_model
-from megatron.training.arguments import parse_args
-from megatron.training.checkpointing import load_checkpoint, save_checkpoint
-from megatron.training.global_vars import set_args
-from megatron.training.training import get_model
-from tests.unit_tests.dist_checkpointing import (
-    TempNamedDir,
-    init_basic_mock_args,
-    init_checkpointing_mock_args,
-)
 from tests.unit_tests.test_utilities import Utils
 from tests.unit_tests.transformer.test_attention import _test_parallel_attention_correctness
 from tests.unit_tests.transformer.test_multi_latent_attention import (
     make_test_packed_seq_params,
     make_test_packed_seq_params_with_padding,
 )
-
-try:
-    import fla
-
-    HAVE_FLA = True
-except ImportError:
-    HAVE_FLA = False
 
 # https://docs.nvidia.com/deeplearning/nccl/user-guide/docs/env.html#nccl-multi-rank-gpu-enable
 # NVLS doesn't support one single GPU to be shared by multiple ranks, so disable this in test
@@ -69,131 +50,16 @@ def _unpack_sequence(x: torch.Tensor, cu_seqlens: torch.Tensor, dim=1) -> list[t
     return unpacked_x
 
 
-try:
-    from causal_conv1d.cpp_functions import causal_conv1d_bwd_function
-except ImportError:
-    HAVE_FUSED_PRE_GDR = False
-else:
-    HAVE_FUSED_PRE_GDR = callable(causal_conv1d_bwd_function)
-
-
-def _make_gdn_config(**overrides):
-    config_kwargs = {
-        "hidden_size": 128,
-        "linear_conv_kernel_dim": 2,
-        "linear_key_head_dim": 32,
-        "linear_value_head_dim": 32,
-        "linear_num_key_heads": 4,
-        "linear_num_value_heads": 8,
-        "num_layers": 1,
-        "normalization": "RMSNorm",
-        "use_cpu_initialization": True,
-        "layernorm_zero_centered_gamma": True,
-        "num_attention_heads": 8,
-        "activation_func": F.silu,
-        "bf16": True,
-        "experimental_attention_variant": "gated_delta_net",
-        "linear_attention_freq": [1],
-        "transformer_impl": "transformer_engine",
-    }
-    config_kwargs.update(overrides)
-    return TransformerConfig(**config_kwargs)
-
-
-def test_gdn_pre_gated_delta_rule_fusion_defaults_to_disabled():
-    config = _make_gdn_config()
-    assert not config.gdn_pre_gated_delta_rule_fusion
-
-
-def test_gdn_pre_gated_delta_rule_fusion_accepts_gdn_variant():
-    config = _make_gdn_config(gdn_pre_gated_delta_rule_fusion=True)
-    assert config.gdn_pre_gated_delta_rule_fusion
-
-
-def test_gdn_pre_gated_delta_rule_fusion_requires_gdn_variant():
-    with pytest.raises(ValueError, match="experimental_attention_variant='gated_delta_net'"):
-        _make_gdn_config(
-            experimental_attention_variant=None,
-            linear_attention_freq=None,
-            gdn_pre_gated_delta_rule_fusion=True,
-        )
-
-
-def test_gdn_conv_pad_alignment_rejects_chunkwise_cp():
-    with pytest.raises(AssertionError, match="gdn_conv_pad_alignment is incompatible"):
-        _make_gdn_config(
-            context_parallel_size=2, linear_cp_mode="chunkwise", gdn_conv_pad_alignment=4096
-        )
-
-
-def test_gdn_chunkwise_cp_head_divisibility_ignores_cp_size():
-    config = _make_gdn_config(
-        tensor_model_parallel_size=2,
-        context_parallel_size=4,
-        linear_cp_mode="chunkwise",
-        linear_num_key_heads=4,
-        linear_num_value_heads=8,
-    )
-    assert config.linear_cp_mode == "chunkwise"
-
-
-def test_fused_pre_gdr_split_batched_recv_send_works():
-    from megatron.core.fusions.fused_pre_gated_delta_rule import _split_batched_recv_send_works
-
-    grouped_work = object()
-    recv_ops, send_ops = _split_batched_recv_send_works([grouped_work], ["recv", "send"])
-    assert recv_ops == (grouped_work,)
-    assert send_ops == ()
-
-    send_work = object()
-    recv_ops, send_ops = _split_batched_recv_send_works([send_work], ["send"])
-    assert recv_ops == ()
-    assert send_ops == (send_work,)
-
-    recv_work = object()
-    send_work = object()
-    recv_ops, send_ops = _split_batched_recv_send_works([recv_work, send_work], ["recv", "send"])
-    assert recv_ops == (recv_work,)
-    assert send_ops == (send_work,)
-
-    with pytest.raises(RuntimeError, match="Expected batch_isend_irecv"):
-        _split_batched_recv_send_works([], ["recv"])
-
-
-def test_gdn_headwise_cp_head_divisibility_includes_cp_size():
-    with pytest.raises(AssertionError, match="linear_head_parallel_size"):
-        _make_gdn_config(
-            tensor_model_parallel_size=2,
-            context_parallel_size=4,
-            linear_cp_mode="headwise",
-            linear_num_key_heads=4,
-            linear_num_value_heads=8,
-        )
-
-
 @pytest.mark.parametrize(
-    ("tp_size", "sp", "cp_size", "linear_cp_mode"),
-    [
-        # cp_size=1: the CP path is inactive, so linear_cp_mode choice is irrelevant.
-        # Cover the "chunkwise" default and skip the "headwise" variants for brevity.
-        (1, False, 1, None),
-        (2, False, 1, None),
-        (2, True, 1, None),
-        # cp_size=2: exercise both CP paths.
-        (1, False, 2, "headwise"),
-        (2, False, 2, "headwise"),
-        (2, True, 2, "headwise"),
-        (1, False, 2, "chunkwise"),
-        (2, False, 2, "chunkwise"),
-        (2, True, 2, "chunkwise"),
-    ],
+    ("tp_size", "sp", "cp_size"),
+    [(1, False, 1), (2, False, 1), (2, True, 1), (1, False, 2), (2, False, 2), (2, True, 2)],
 )
 @pytest.mark.skipif(not HAVE_FLA, reason="FLA is not installed.")
 @pytest.mark.internal
 class TestGatedDeltaNet:
 
     @pytest.fixture(scope='function', autouse=True)
-    def setup_method(self, tp_size, sp, cp_size, linear_cp_mode):
+    def setup_method(self, tp_size, sp, cp_size):
         # Initialize parallel and random seed
         Utils.initialize_model_parallel(
             tensor_model_parallel_size=tp_size,
@@ -204,18 +70,6 @@ class TestGatedDeltaNet:
         self.tp_size = tp_size
         self.cp_size = cp_size
         self.sp_size = tp_size if sp else 1
-        self.linear_cp_mode = linear_cp_mode
-        if self.linear_cp_mode == "headwise":
-            self.cp_size_chunkwise = 1
-            self.cp_size_headwise = self.cp_size
-        elif self.linear_cp_mode == "chunkwise":
-            self.cp_size_chunkwise = self.cp_size
-            self.cp_size_headwise = 1
-        elif self.cp_size == 1:
-            self.cp_size_chunkwise = 1
-            self.cp_size_headwise = 1
-        else:
-            raise ValueError(f"Invalid linear CP mode: {self.linear_cp_mode}")
 
         # Get TP and CP process groups from device mesh
         tp_group = parallel_state.get_tensor_model_parallel_group()
@@ -243,16 +97,13 @@ class TestGatedDeltaNet:
             context_parallel_size=cp_size,
             experimental_attention_variant="gated_delta_net",
             linear_attention_freq=[1],
-            linear_cp_mode=self.linear_cp_mode,
             transformer_impl="transformer_engine",
         )
-        gdn_submodules = get_experimental_attention_variant_module_spec(
-            config=self.transformer_config
-        ).submodules
+        gdn_spec = get_experimental_attention_variant_module_spec(config=self.transformer_config)
 
-        self.gdn = GatedDeltaNet(
+        self.gdn = gdn_spec.module(
             self.transformer_config,
-            submodules=gdn_submodules,
+            submodules=gdn_spec.submodules,
             layer_number=1,
             bias=False,
             conv_bias=False,
@@ -269,7 +120,7 @@ class TestGatedDeltaNet:
     def test_gpu_forward(self):
         gdn = self.gdn
 
-        micro_batch_size = 1 if self.linear_cp_mode == "chunkwise" and self.cp_size > 1 else 2
+        micro_batch_size = 2
         seq_length = 64
         hidden_states = torch.ones(
             (seq_length // self.sp_size // self.cp_size, micro_batch_size, gdn.config.hidden_size),
@@ -295,95 +146,156 @@ class TestGatedDeltaNet:
             output.dtype == hidden_states.dtype
         ), f"Output dtype {output.dtype=} mismatch with {hidden_states.dtype=}"
 
-    @pytest.mark.flaky_in_dev  # Issue #5473
-    def test_selective_recompute_gdn(self):
-        """Whole-module 'gdn' recompute must match the non-recompute forward and gradients.
+    def test_selective_recompute_norm_out(self):
+        tp_group = parallel_state.get_tensor_model_parallel_group()
+        cp_group = parallel_state.get_context_parallel_group()
+        pg_collection = ProcessGroupCollection(tp=tp_group, cp=cp_group)
 
-        The same module/input is run twice (recompute off, then on); the forward output and
-        all parameter / input gradients must agree within a tight tolerance (rtol/atol=1e-4).
-        The recompute path is run-to-run deterministic on these kernels (empirically bitwise),
-        so a tolerance well below the bf16 floor is expected to hold.
-        """
-        gdn = self.gdn
-        gdn.train()
+        def build_gdn(config):
+            gdn_spec = get_experimental_attention_variant_module_spec(config=config)
+            gdn = gdn_spec.module(
+                config,
+                submodules=gdn_spec.submodules,
+                layer_number=1,
+                bias=False,
+                conv_bias=False,
+                conv_init=1.0,
+                use_qk_l2norm=True,
+                A_init_range=(1, 16),
+                pg_collection=pg_collection,
+            )
+            return gdn.cuda().bfloat16()
 
-        micro_batch_size = 1 if self.linear_cp_mode == "chunkwise" and self.cp_size > 1 else 2
+        def run(gdn, hidden_states):
+            output, _ = gdn(hidden_states, None)
+            output.float().sum().backward()
+            grads = {
+                name: param.grad.detach()
+                for name, param in gdn.named_parameters()
+                if param.grad is not None
+            }
+            input_grad = hidden_states.grad.detach().clone()
+            return output.detach(), grads, input_grad
+
+        micro_batch_size = 2
         seq_length = 64
-        torch.manual_seed(1234)
+        base_config = copy.deepcopy(self.transformer_config)
+        rec_config = copy.deepcopy(self.transformer_config)
+        rec_config.recompute_granularity = "selective"
+        rec_config.recompute_modules = ["gdn_norm_out"]
+
+        model_parallel_cuda_manual_seed(42)
+        torch.manual_seed(42)
+        hidden_states = torch.randn(
+            (
+                seq_length // self.sp_size // self.cp_size,
+                micro_batch_size,
+                self.gdn.config.hidden_size,
+            ),
+            device=torch.cuda.current_device(),
+            dtype=torch.bfloat16,
+            requires_grad=True,
+        )
+
+        # --- Baseline (no recompute) ---
+        model_parallel_cuda_manual_seed(42)
+        torch.manual_seed(42)
+        base_gdn = build_gdn(base_config)
+        assert base_gdn.recompute_norm_out is False
+        base_output, base_grads, base_input_grad = run(base_gdn, hidden_states)
+        hidden_states.grad = None
+        assert base_gdn.norm_out_checkpoint is None
+        del base_gdn
+        torch.cuda.empty_cache()
+
+        # --- Recompute ---
+        model_parallel_cuda_manual_seed(42)
+        torch.manual_seed(42)
+        rec_gdn = build_gdn(rec_config)
+        assert rec_gdn.recompute_norm_out is True
+        rec_output, rec_grads, rec_input_grad = run(rec_gdn, hidden_states)
+        assert rec_gdn.norm_out_checkpoint is not None
+
+        rank = torch.distributed.get_rank()
+        assert torch.equal(rec_output, base_output), f"Output not identical ({rank=})"
+        assert torch.equal(rec_input_grad, base_input_grad), f"Input grad not identical ({rank=})"
+        assert set(rec_grads.keys()) == set(base_grads.keys())
+        for name in base_grads:
+            assert torch.equal(
+                rec_grads[name], base_grads[name]
+            ), f"Grad not identical for {name} ({rank=})"
+
+    def test_deterministic_mode(self):
+        tp_group = parallel_state.get_tensor_model_parallel_group()
+        cp_group = parallel_state.get_context_parallel_group()
+        pg_collection = ProcessGroupCollection(tp=tp_group, cp=cp_group)
+
+        det_config = copy.deepcopy(self.transformer_config)
+        det_config.deterministic_mode = True
+
+        gdn_spec = get_experimental_attention_variant_module_spec(config=det_config)
+
+        model_parallel_cuda_manual_seed(42)
+        torch.manual_seed(42)
+        gdn = (
+            gdn_spec.module(
+                det_config,
+                submodules=gdn_spec.submodules,
+                layer_number=1,
+                bias=False,
+                conv_bias=False,
+                conv_init=1.0,
+                use_qk_l2norm=True,
+                A_init_range=(1, 16),
+                pg_collection=pg_collection,
+            )
+            .cuda()
+            .bfloat16()
+        )
+
+        # deterministic_mode must select the torch-native kernel, not FLA.
+        assert isinstance(gdn, GatedDeltaNet)
+        assert gdn.gated_delta_rule is torch_chunk_gated_delta_rule
+
+        micro_batch_size = 2
+        seq_length = 64
+        torch.manual_seed(0)
         base_input = torch.randn(
             (seq_length // self.sp_size // self.cp_size, micro_batch_size, gdn.config.hidden_size),
             device=torch.cuda.current_device(),
             dtype=torch.bfloat16,
         )
 
-        def run(recompute):
-            gdn.recompute_gdn = recompute
-            gdn.zero_grad(set_to_none=True)
-            hidden_states = base_input.clone().detach().requires_grad_(True)
-            output, _ = gdn(hidden_states, None)
-            output.float().square().mean().backward()
-            param_grads = {
+        def run(module):
+            hidden_states = base_input.clone().requires_grad_(True)
+            output, _ = module(hidden_states, None)
+            output.float().sum().backward()
+            grads = {
                 name: param.grad.detach().clone()
-                for name, param in gdn.named_parameters()
+                for name, param in module.named_parameters()
                 if param.grad is not None
             }
-            return output.detach().clone(), hidden_states.grad.detach().clone(), param_grads
+            module.zero_grad(set_to_none=True)
+            return output.detach().clone(), grads, hidden_states.grad.detach().clone()
 
-        try:
-            out_ref, dinput_ref, pgrad_ref = run(recompute=False)
-            out_rc, dinput_rc, pgrad_rc = run(recompute=True)
-        finally:
-            gdn.recompute_gdn = False
+        out1, grads1, input_grad1 = run(gdn)
+        out2, grads2, input_grad2 = run(gdn)
 
-        torch.testing.assert_close(out_rc, out_ref, rtol=1e-4, atol=1e-4)
-        torch.testing.assert_close(dinput_rc, dinput_ref, rtol=1e-4, atol=1e-4)
-        assert pgrad_ref.keys() == pgrad_rc.keys(), "recompute changed the set of grad params"
-        assert len(pgrad_ref) > 0, "expected at least one parameter gradient"
-        for name in pgrad_ref:
-            torch.testing.assert_close(
-                pgrad_rc[name],
-                pgrad_ref[name],
-                rtol=1e-4,
-                atol=1e-4,
-                msg=lambda m, n=name: f"gradient mismatch for parameter '{n}': {m}",
-            )
+        rank = torch.distributed.get_rank()
+        assert torch.equal(out1, out2), f"Output not reproducible ({rank=})"
+        assert torch.equal(input_grad1, input_grad2), f"Input grad not reproducible ({rank=})"
+        assert set(grads1.keys()) == set(grads2.keys())
+        for name in grads1:
+            assert torch.equal(
+                grads1[name], grads2[name]
+            ), f"Grad not reproducible for {name} ({rank=})"
 
-    def test_gpu_forward_rejects_sbhd_chunkwise_cp_batch_gt_one(self):
-        if not (self.linear_cp_mode == "chunkwise" and self.cp_size > 1):
-            pytest.skip("Only chunkwise CP with CP>1 uses the FLA CP batch guard.")
-
+    def test_module_construction(self):
         gdn = self.gdn
-
-        micro_batch_size = 2
-        seq_length = 64
-        hidden_states = torch.ones(
-            (seq_length // self.sp_size // self.cp_size, micro_batch_size, gdn.config.hidden_size),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-
-        with pytest.raises(ValueError, match="requires micro_batch_size == 1"):
-            gdn(hidden_states, None)
-
-    def test_gpu_forward_rejects_sbhd_conv_padding(self):
-        gdn = self.gdn
-        gdn.config.gdn_conv_pad_alignment = 4096
-
-        micro_batch_size = 1 if self.linear_cp_mode == "chunkwise" and self.cp_size > 1 else 2
-        seq_length = 64
-        hidden_states = torch.ones(
-            (seq_length // self.sp_size // self.cp_size, micro_batch_size, gdn.config.hidden_size),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-
-        expected_error = (
-            "incompatible with GDN chunkwise CP"
-            if self.linear_cp_mode == "chunkwise" and self.cp_size > 1
-            else "only supported with packed sequence"
-        )
-        with pytest.raises(ValueError, match=expected_error):
-            gdn(hidden_states, None)
+        assert isinstance(gdn, GatedDeltaNet)
+        assert gdn.in_proj_dim == 2 * gdn.qk_dim + 2 * gdn.v_dim + 2 * gdn.num_value_heads
+        assert gdn.A_log.shape == (gdn.num_value_heads // self.tp_size,)
+        assert gdn.dt_bias.shape == (gdn.num_value_heads // self.tp_size,)
 
     def test_jit_compiled_helpers(self):
         import torch._dynamo
@@ -392,138 +304,60 @@ class TestGatedDeltaNet:
         batch = 2
         seq_len = 16
 
-        num_v_heads_local = gdn.num_value_heads // gdn.tp_size // self.cp_size_headwise
+        device = torch.cuda.current_device()
+        num_v_heads_local = gdn.num_value_heads // gdn.tp_size // gdn.cp_size
+        qk_dim_local = gdn.qk_dim_local_tp // gdn.cp_size
+        v_dim_local = gdn.v_dim_local_tp // gdn.cp_size
 
-        qkv_last_dim = (2 * gdn.qk_dim_local_tp + gdn.v_dim_local_tp) // self.cp_size_headwise
         qkv = torch.randn(
-            batch, seq_len, qkv_last_dim, device=torch.cuda.current_device(), dtype=torch.bfloat16
+            batch, seq_len, 2 * qk_dim_local + v_dim_local, device=device, dtype=torch.bfloat16
         )
         gate = torch.randn(
             batch,
             seq_len,
             num_v_heads_local,
             gdn.value_head_dim,
-            device=torch.cuda.current_device(),
+            device=device,
             dtype=torch.bfloat16,
         )
-        beta = torch.randn(
-            batch,
-            seq_len,
-            num_v_heads_local,
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        alpha = torch.randn(
-            batch,
-            seq_len,
-            num_v_heads_local,
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
+        gate_feats = (
+            torch.randn(batch, seq_len, num_v_heads_local, device=device, dtype=torch.bfloat16),
+            torch.randn(batch, seq_len, num_v_heads_local, device=device, dtype=torch.bfloat16),
+        )  # beta, alpha
+        A_log_mock = torch.randn(num_v_heads_local, device=device, dtype=torch.bfloat16)
+        dt_bias_mock = torch.randn(num_v_heads_local, device=device, dtype=torch.bfloat16)
+        expected_keys = {"q", "k", "v", "g", "beta"}
 
         # Disable dynamo so coverage.py can trace through the method bodies,
         # which are normally wrapped by @jit_fuser (torch.compile).
         with torch._dynamo.config.patch(disable=True):
-            query, key, value, gate_out, beta_out, alpha_out = (
-                gdn._prepare_qkv_for_gated_delta_rule(
-                    qkv, gate, beta, alpha, batch, seq_len, cp_size_headwise=self.cp_size_headwise
-                )
+            kernel_inputs = gdn._prepare_input_for_gated_delta_rule(
+                qkv, gate, A_log_mock, dt_bias_mock, batch, seq_len, *gate_feats
             )
 
+        # The output gate (z) rides along under "gate" and is popped by forward before
+        # the kernel call; everything else is passed straight through as kernel kwargs.
+        gate_out = kernel_inputs.pop("gate")
+        assert set(kernel_inputs) == expected_keys
+
+        query, key, value, g = (kernel_inputs[k] for k in ("q", "k", "v", "g"))
         assert query.shape == (batch, seq_len, num_v_heads_local, gdn.key_head_dim)
         assert key.shape == (batch, seq_len, num_v_heads_local, gdn.key_head_dim)
         assert value.shape == (batch, seq_len, num_v_heads_local, gdn.value_head_dim)
-        assert query.is_contiguous()
-        assert key.is_contiguous()
-        assert value.is_contiguous()
+        assert gate_out.shape == (batch, seq_len, num_v_heads_local, gdn.value_head_dim)
+        for t in (query, key, value, gate_out, *kernel_inputs.values()):
+            assert t.is_contiguous()
 
-        A_log_mock = torch.randn(
-            num_v_heads_local, device=torch.cuda.current_device(), dtype=torch.bfloat16
-        )
-        dt_bias_mock = torch.randn(
-            num_v_heads_local, device=torch.cuda.current_device(), dtype=torch.bfloat16
-        )
-
-        with torch._dynamo.config.patch(disable=True):
-            g, beta_sig = gdn._compute_g_and_beta(A_log_mock, dt_bias_mock, alpha, beta)
-
-        assert g.dtype == torch.float32
-        assert g.shape == alpha.shape
-        assert beta_sig.shape == beta.shape
-
-    def test_fused_pre_gated_delta_rule_headwise_cp_uses_cp_local_parameters(self):
-        if not HAVE_FUSED_PRE_GDR:
-            pytest.skip("causal-conv1d fused backward is not installed.")
-        if not (self.linear_cp_mode == "headwise" and self.cp_size > 1):
-            pytest.skip("Only headwise CP with CP>1 needs CP-local fused pre-GDR params.")
-
-        gdn = self.gdn
-        batch = 2
-        seq_len = 16
-        qk_channels = gdn.qk_dim_local_tp // self.cp_size_headwise
-        v_channels = gdn.v_dim_local_tp // self.cp_size_headwise
-        num_key_heads = qk_channels // gdn.key_head_dim
-        num_value_heads = v_channels // gdn.value_head_dim
-        qkvzba_dim = 2 * qk_channels + 2 * v_channels + 2 * num_value_heads
-        qkvzba = torch.randn(
-            seq_len, batch, qkvzba_dim, device=torch.cuda.current_device(), dtype=torch.bfloat16
-        )
-        captured = {}
-
-        def fake_fused_streamed_pre_gated_delta_rule(
-            qkvzba_arg,
-            conv1d_weight,
-            conv1d_bias,
-            A_log,
-            dt_bias,
-            *,
-            num_key_heads,
-            num_value_heads,
-            **kwargs,
-        ):
-            captured.update(
-                {
-                    "qkvzba": qkvzba_arg,
-                    "conv1d_weight": conv1d_weight,
-                    "conv1d_bias": conv1d_bias,
-                    "A_log": A_log,
-                    "dt_bias": dt_bias,
-                    "num_key_heads": num_key_heads,
-                    "num_value_heads": num_value_heads,
-                    "cp_group": kwargs["cp_group"],
-                }
-            )
-            return tuple(torch.empty(0, device=qkvzba_arg.device) for _ in range(6))
-
-        with mock.patch(
-            "megatron.core.fusions.fused_pre_gated_delta_rule."
-            "fused_streamed_pre_gated_delta_rule",
-            side_effect=fake_fused_streamed_pre_gated_delta_rule,
-        ):
-            gdn._fused_streamed_pre_gated_delta_rule(
-                qkvzba,
-                cp_size_headwise=self.cp_size_headwise,
-                cp_group_headwise=gdn.pg_collection.cp,
-            )
-
-        assert captured["qkvzba"] is qkvzba
-        assert captured["conv1d_weight"].shape == (
-            2 * qk_channels + v_channels,
-            1,
-            gdn.conv_kernel_dim,
-        )
-        assert captured["conv1d_bias"] is None
-        assert captured["A_log"].shape == (num_value_heads,)
-        assert captured["dt_bias"].shape == (num_value_heads,)
-        assert captured["num_key_heads"] == num_key_heads
-        assert captured["num_value_heads"] == num_value_heads
-        assert captured["cp_group"] is None
+        # Per-head decay and write strength beta
+        beta = kernel_inputs["beta"]
+        assert g.shape == (batch, seq_len, num_v_heads_local)
+        assert beta.shape == (batch, seq_len, num_v_heads_local)
+        assert (g <= 0).all()
+        assert (beta >= 0).all() and (beta <= 1).all()
 
     def test_gpu_forward_thd_correctness(self):
         if self.sp_size > 1:
             pytest.skip("Sequence parallel is not supported for this test case.")
-        if self.cp_size > 1 and self.linear_cp_mode == "chunkwise":
-            pytest.skip("Chunkwise CP is not supported for this test case.")
 
         atol, rtol = 3e-4, 3e-4
 
@@ -567,8 +401,6 @@ class TestGatedDeltaNet:
     def test_gpu_forward_thd_padding_correctness(self):
         if self.sp_size > 1:
             pytest.skip("Sequence parallel is not supported for this test case.")
-        if self.cp_size > 1 and self.linear_cp_mode == "chunkwise":
-            pytest.skip("Chunkwise CP is not supported for this test case.")
 
         atol, rtol = 3e-4, 3e-4
         sequence_length = 32
@@ -611,1015 +443,17 @@ class TestGatedDeltaNet:
         )
         assert output_thd_no_padding.shape == output_thd_padded.shape
 
-        # C) explicit causal-conv padding is only applied to packed inputs and
-        # should not affect the original unpadded token outputs.
-        self.gdn.config.gdn_conv_pad_alignment = 48
-        output_thd_conv_pad, _ = self.gdn(
-            hidden_states_thd, None, packed_seq_params=no_padding_params
-        )
-        self.gdn.config.gdn_conv_pad_alignment = None
-        assert output_thd_conv_pad.shape == output_thd_no_padding.shape
-        torch.testing.assert_close(
-            output_thd_conv_pad,
-            output_thd_no_padding,
-            atol=atol,
-            rtol=rtol,
-            msg=lambda msg: f"THD conv-padded output mismatch ({rank=}): {msg}",
-        )
-
-        # D) padded mismatch branch: if *_padded[-1] mismatches total_sequence_length, should raise.
+        # C) padded mismatch branch: if *_padded[-1] mismatches total_sequence_length, should raise.
         padded_mismatch_params = make_test_packed_seq_params_with_padding(
             cu_seqlens=[0, 30, 60, 90, 120], cu_seqlens_padded=[0, 32, 64, 96, 126]
         )
         with pytest.raises(ValueError, match="does not match"):
             self.gdn(hidden_states_thd, None, packed_seq_params=padded_mismatch_params)
 
-        # E) actual mismatch branch without *_padded: should raise.
+        # D) actual mismatch branch without *_padded: should raise.
         actual_mismatch_params = make_test_packed_seq_params(cu_seqlens=[0, 32, 64, 96, 129])
         with pytest.raises(ValueError, match="does not match"):
             self.gdn(hidden_states_thd, None, packed_seq_params=actual_mismatch_params)
-
-
-@pytest.mark.skipif(not HAVE_FLA, reason="FLA is not installed.")
-@pytest.mark.skipif(not HAVE_FUSED_PRE_GDR, reason="causal-conv1d fused backward is not installed.")
-@pytest.mark.internal
-class TestFusedPreGatedDeltaRule:
-
-    @pytest.fixture(scope='function', autouse=True)
-    def setup_method(self):
-        Utils.initialize_model_parallel(
-            tensor_model_parallel_size=1, pipeline_model_parallel_size=1, context_parallel_size=1
-        )
-        model_parallel_cuda_manual_seed(123)
-
-        tp_group = parallel_state.get_tensor_model_parallel_group()
-        cp_group = parallel_state.get_context_parallel_group()
-        self.pg_collection = ProcessGroupCollection(tp=tp_group, cp=cp_group)
-
-        self.unfused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=False)
-        self.fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True, deterministic_mode=False
-        )
-        self.fused_gdn.load_state_dict(self.unfused_gdn.state_dict())
-
-    def teardown_method(self):
-        Utils.destroy_model_parallel()
-
-    def _build_gdn(
-        self,
-        gdn_pre_gated_delta_rule_fusion: bool,
-        *,
-        deterministic_mode: bool = True,
-        conv_kernel_dim: int = 2,
-        hidden_size: int = 256,
-        key_head_dim: int = 64,
-        value_head_dim: int = 64,
-        num_key_heads: int = 4,
-        num_value_heads: int = 8,
-    ):
-        transformer_config = TransformerConfig(
-            hidden_size=hidden_size,
-            linear_conv_kernel_dim=conv_kernel_dim,
-            linear_key_head_dim=key_head_dim,
-            linear_value_head_dim=value_head_dim,
-            linear_num_key_heads=num_key_heads,
-            linear_num_value_heads=num_value_heads,
-            num_layers=1,
-            normalization="RMSNorm",
-            use_cpu_initialization=True,
-            layernorm_zero_centered_gamma=True,
-            num_attention_heads=num_value_heads,
-            activation_func=F.silu,
-            bf16=True,
-            tensor_model_parallel_size=1,
-            context_parallel_size=1,
-            experimental_attention_variant="gated_delta_net",
-            linear_attention_freq=[1],
-            transformer_impl="transformer_engine",
-            deterministic_mode=deterministic_mode,
-            gdn_pre_gated_delta_rule_fusion=gdn_pre_gated_delta_rule_fusion,
-        )
-        gdn_submodules = get_experimental_attention_variant_module_spec(
-            config=transformer_config
-        ).submodules
-        gdn = GatedDeltaNet(
-            transformer_config,
-            submodules=gdn_submodules,
-            layer_number=1,
-            bias=False,
-            conv_bias=False,
-            conv_init=1.0,
-            use_qk_l2norm=True,
-            A_init_range=(1, 16),
-            pg_collection=self.pg_collection,
-        )
-        return gdn.cuda().bfloat16()
-
-    def _packed_pre_gated_delta_rule_reference(self, gdn, qkvzba, cu_seqlens):
-        """Run the unfused pre-GDR path independently on each packed sequence."""
-
-        segment_outputs = [[] for _ in range(6)]
-        for start, end in zip(cu_seqlens[:-1].tolist(), cu_seqlens[1:].tolist()):
-            outputs = gdn.pre_gated_delta_rule(
-                qkvzba[start:end],
-                batch=1,
-                seq_len=end - start,
-                cp_size_headwise=gdn.cp_size,
-                cp_group_headwise=gdn.pg_collection.cp,
-            )
-            for output_list, output in zip(segment_outputs, outputs):
-                output_list.append(output)
-        return tuple(torch.cat(outputs, dim=1) for outputs in segment_outputs)
-
-    def _assert_pre_gated_delta_rule_outputs_close(
-        self, fused_outputs, unfused_outputs, *, atol: float, rtol: float, output_tolerances=None
-    ):
-        """Compare named pre-GDR outputs with optional per-output tolerances."""
-
-        output_names = ("query", "key", "value", "gate", "beta", "g")
-        output_tolerances = output_tolerances or {}
-        for name, fused, unfused in zip(output_names, fused_outputs, unfused_outputs):
-            output_atol, output_rtol = output_tolerances.get(name, (atol, rtol))
-            torch.testing.assert_close(
-                fused,
-                unfused,
-                atol=output_atol,
-                rtol=output_rtol,
-                msg=lambda msg, output_name=name: f"{output_name} mismatch: {msg}",
-            )
-
-    def test_fused_and_unfused_forward_match(self):
-        hidden_states = torch.randn(
-            (32, 2, self.unfused_gdn.config.hidden_size),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-
-        with torch.no_grad():
-            unfused_output, unfused_bias = self.unfused_gdn(hidden_states, None)
-            fused_output, fused_bias = self.fused_gdn(hidden_states, None)
-
-        torch.testing.assert_close(fused_output, unfused_output, atol=1e-3, rtol=1e-3)
-        assert fused_bias == unfused_bias
-
-    def test_fused_and_unfused_forward_thd_match(self):
-        unfused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=False, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn.load_state_dict(unfused_gdn.state_dict())
-
-        hidden_states = torch.randn(
-            (32, 1, unfused_gdn.config.hidden_size),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        cu_seqlens = torch.tensor(
-            [0, 1, 4, 11, 32], device=torch.cuda.current_device(), dtype=torch.int32
-        )
-        packed_seq_params = PackedSeqParams(
-            qkv_format='thd',
-            cu_seqlens_q=cu_seqlens,
-            cu_seqlens_kv=cu_seqlens,
-            max_seqlen_q=21,
-            max_seqlen_kv=21,
-            total_tokens=hidden_states.shape[0],
-        )
-        assert packed_seq_params.seq_idx is not None
-
-        with torch.no_grad():
-            unfused_output, unfused_bias = unfused_gdn(
-                hidden_states, None, packed_seq_params=packed_seq_params
-            )
-            fused_output, fused_bias = fused_gdn(
-                hidden_states, None, packed_seq_params=packed_seq_params
-            )
-
-        torch.testing.assert_close(fused_output, unfused_output, atol=2e-3, rtol=2e-3)
-        assert fused_bias == unfused_bias
-
-    def test_fused_and_unfused_forward_thd_padding_match(self):
-        unfused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=False, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn.load_state_dict(unfused_gdn.state_dict())
-
-        hidden_states = torch.randn(
-            (12, 1, unfused_gdn.config.hidden_size),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        cu_seqlens = torch.tensor(
-            [0, 1, 4, 9], device=torch.cuda.current_device(), dtype=torch.int32
-        )
-        cu_seqlens_padded = torch.tensor(
-            [0, 2, 6, 12], device=torch.cuda.current_device(), dtype=torch.int32
-        )
-        packed_seq_params = PackedSeqParams(
-            qkv_format='thd',
-            cu_seqlens_q=cu_seqlens,
-            cu_seqlens_kv=cu_seqlens,
-            cu_seqlens_q_padded=cu_seqlens_padded,
-            cu_seqlens_kv_padded=cu_seqlens_padded,
-            max_seqlen_q=6,
-            max_seqlen_kv=6,
-            total_tokens=hidden_states.shape[0],
-        )
-        assert packed_seq_params.seq_idx is not None
-
-        with torch.no_grad():
-            unfused_output, unfused_bias = unfused_gdn(
-                hidden_states, None, packed_seq_params=packed_seq_params
-            )
-            fused_output, fused_bias = fused_gdn(
-                hidden_states, None, packed_seq_params=packed_seq_params
-            )
-
-        torch.testing.assert_close(fused_output, unfused_output, atol=2e-3, rtol=2e-3)
-        assert fused_bias == unfused_bias
-
-    def test_fused_and_unfused_backward_thd_padding_match(self):
-        unfused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=False, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn.load_state_dict(unfused_gdn.state_dict())
-
-        real_seq_len = 9
-        padded_seq_len = 16
-        hidden_states = torch.randn(
-            (padded_seq_len, 1, unfused_gdn.config.hidden_size),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        hidden_states_unfused = hidden_states.detach().clone().requires_grad_(True)
-        hidden_states_fused = hidden_states.detach().clone().requires_grad_(True)
-        cu_seqlens = torch.tensor(
-            [0, 5, real_seq_len, padded_seq_len],
-            device=torch.cuda.current_device(),
-            dtype=torch.int32,
-        )
-        packed_seq_params = PackedSeqParams(
-            qkv_format='thd',
-            cu_seqlens_q=cu_seqlens,
-            cu_seqlens_kv=cu_seqlens,
-            cu_seqlens_q_padded=cu_seqlens,
-            cu_seqlens_kv_padded=cu_seqlens,
-            max_seqlen_q=padded_seq_len - real_seq_len,
-            max_seqlen_kv=padded_seq_len - real_seq_len,
-            total_tokens=hidden_states.shape[0],
-        )
-        assert packed_seq_params.seq_idx is not None
-
-        unfused_gdn.zero_grad(set_to_none=True)
-        fused_gdn.zero_grad(set_to_none=True)
-        unfused_output, _ = unfused_gdn(
-            hidden_states_unfused, None, packed_seq_params=packed_seq_params
-        )
-        fused_output, _ = fused_gdn(hidden_states_fused, None, packed_seq_params=packed_seq_params)
-        grad_output = torch.randn_like(unfused_output.float())
-        grad_output[real_seq_len:] = 0.0
-
-        unfused_loss = (unfused_output.float() * grad_output).sum()
-        fused_loss = (fused_output.float() * grad_output).sum()
-        unfused_loss.backward()
-        fused_loss.backward()
-
-        torch.testing.assert_close(
-            hidden_states_fused.grad, hidden_states_unfused.grad, atol=3e-2, rtol=3e-2
-        )
-        for param_name in ("in_proj.weight", "conv1d.weight", "A_log", "dt_bias"):
-            fused_param = dict(fused_gdn.named_parameters())[param_name]
-            unfused_param = dict(unfused_gdn.named_parameters())[param_name]
-            torch.testing.assert_close(
-                fused_param.grad,
-                unfused_param.grad,
-                atol=3e-2,
-                rtol=3e-2,
-                msg=lambda msg, name=param_name: f"{name} grad mismatch: {msg}",
-            )
-
-    def test_fused_and_unfused_pre_gated_delta_rule_match(self):
-        batch = 2
-        seq_len = 32
-        hidden_states = torch.randn(
-            (seq_len, batch, self.unfused_gdn.config.hidden_size),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-
-        with torch.no_grad():
-            qkvzba, _ = self.unfused_gdn.in_proj(hidden_states)
-            unfused_outputs = self.unfused_gdn.pre_gated_delta_rule(
-                qkvzba, batch, seq_len, self.unfused_gdn.cp_size, self.unfused_gdn.pg_collection.cp
-            )
-            fused_outputs = self.fused_gdn._fused_streamed_pre_gated_delta_rule(qkvzba)
-
-        self._assert_pre_gated_delta_rule_outputs_close(
-            fused_outputs,
-            unfused_outputs,
-            atol=1e-3,
-            rtol=1e-3,
-            output_tolerances={"g": (1e-3, 3e-3)},
-        )
-
-    def test_fused_and_unfused_pre_gated_delta_rule_backward_match(self):
-        reference_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=False, deterministic_mode=True, conv_kernel_dim=4
-        )
-        fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn.load_state_dict(reference_gdn.state_dict())
-
-        batch = 2
-        seq_len = 32
-        qkvzba = torch.randn(
-            (seq_len, batch, reference_gdn.in_proj_dim),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        qkvzba_unfused = qkvzba.detach().clone().requires_grad_(True)
-        qkvzba_fused = qkvzba.detach().clone().requires_grad_(True)
-
-        reference_gdn.zero_grad(set_to_none=True)
-        fused_gdn.zero_grad(set_to_none=True)
-
-        unfused_outputs = reference_gdn.pre_gated_delta_rule(
-            qkvzba_unfused, batch, seq_len, reference_gdn.cp_size, reference_gdn.pg_collection.cp
-        )
-        fused_outputs = fused_gdn._fused_streamed_pre_gated_delta_rule(qkvzba_fused)
-        grad_outputs = [torch.randn_like(output.float()) for output in unfused_outputs]
-
-        unfused_loss = sum(
-            (output.float() * grad).sum() for output, grad in zip(unfused_outputs, grad_outputs)
-        )
-        fused_loss = sum(
-            (output.float() * grad).sum() for output, grad in zip(fused_outputs, grad_outputs)
-        )
-        unfused_loss.backward()
-        fused_loss.backward()
-
-        torch.testing.assert_close(qkvzba_fused.grad, qkvzba_unfused.grad, atol=3e-2, rtol=3e-2)
-        torch.testing.assert_close(
-            fused_gdn.conv1d.weight.grad, reference_gdn.conv1d.weight.grad, atol=3e-2, rtol=3e-2
-        )
-        torch.testing.assert_close(
-            fused_gdn.A_log.grad, reference_gdn.A_log.grad, atol=3e-2, rtol=3e-2
-        )
-        torch.testing.assert_close(
-            fused_gdn.dt_bias.grad, reference_gdn.dt_bias.grad, atol=3e-2, rtol=3e-2
-        )
-
-    def test_fused_and_unfused_packed_pre_gated_delta_rule_forward_match(self):
-        reference_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=False, deterministic_mode=True, conv_kernel_dim=4
-        )
-        fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn.load_state_dict(reference_gdn.state_dict())
-
-        batch = 1
-        cu_seqlens = torch.tensor(
-            [0, 1, 4, 6, 11], device=torch.cuda.current_device(), dtype=torch.int32
-        )
-        seq_len = cu_seqlens[-1].item()
-        qkvzba = torch.randn(
-            (seq_len, batch, reference_gdn.in_proj_dim),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-
-        with torch.no_grad():
-            unfused_outputs = self._packed_pre_gated_delta_rule_reference(
-                reference_gdn, qkvzba, cu_seqlens
-            )
-            fused_outputs = fused_gdn._fused_streamed_pre_gated_delta_rule(
-                qkvzba, cu_seqlens_q=cu_seqlens
-            )
-
-        self._assert_pre_gated_delta_rule_outputs_close(
-            fused_outputs, unfused_outputs, atol=2e-3, rtol=2e-3
-        )
-
-    def test_fused_packed_g_softplus_matches_torch_for_large_alpha(self):
-        fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True, deterministic_mode=False, conv_kernel_dim=4
-        )
-        device = torch.cuda.current_device()
-        batch = 1
-        cu_seqlens = torch.tensor([0, 3, 8], device=device, dtype=torch.int32)
-        seq_len = cu_seqlens[-1].item()
-        num_value_heads = fused_gdn.num_v_heads_local_tp
-        beta_channel_offset = 2 * fused_gdn.qk_dim_local_tp + 2 * fused_gdn.v_dim_local_tp
-        alpha_channel_offset = beta_channel_offset + num_value_heads
-
-        qkvzba = torch.zeros(
-            (seq_len, batch, fused_gdn.in_proj_dim),
-            device=device,
-            dtype=torch.bfloat16,
-            requires_grad=True,
-        )
-        with torch.no_grad():
-            qkvzba[:, :, alpha_channel_offset : alpha_channel_offset + num_value_heads] = 100.0
-            fused_gdn.A_log.zero_()
-            fused_gdn.dt_bias.zero_()
-        fused_gdn.zero_grad(set_to_none=True)
-
-        *_, g = fused_gdn._fused_streamed_pre_gated_delta_rule(qkvzba, cu_seqlens_q=cu_seqlens)
-
-        alpha = qkvzba[
-            :, :, alpha_channel_offset : alpha_channel_offset + num_value_heads
-        ].transpose(0, 1)
-        expected_g = -torch.exp(fused_gdn.A_log.float()).view(1, 1, -1) * F.softplus(
-            alpha.float() + fused_gdn.dt_bias.float().view(1, 1, -1)
-        )
-        assert torch.isfinite(g).all()
-        torch.testing.assert_close(g.float(), expected_g, atol=0.0, rtol=0.0)
-
-        g.float().sum().backward()
-        expected_alpha_grad = -torch.ones_like(alpha.float())
-        expected_A_log_grad = expected_g.sum(dim=(0, 1))
-        expected_dt_bias_grad = -torch.full_like(fused_gdn.dt_bias.float(), seq_len * batch)
-        alpha_grad = qkvzba.grad[
-            :, :, alpha_channel_offset : alpha_channel_offset + num_value_heads
-        ].transpose(0, 1)
-
-        assert torch.isfinite(alpha_grad).all()
-        assert torch.isfinite(fused_gdn.A_log.grad).all()
-        assert torch.isfinite(fused_gdn.dt_bias.grad).all()
-        torch.testing.assert_close(alpha_grad.float(), expected_alpha_grad, atol=0.0, rtol=0.0)
-        torch.testing.assert_close(
-            fused_gdn.A_log.grad.float(), expected_A_log_grad, atol=0.0, rtol=0.0
-        )
-        torch.testing.assert_close(
-            fused_gdn.dt_bias.grad.float(), expected_dt_bias_grad, atol=0.0, rtol=0.0
-        )
-
-    def test_fused_and_unfused_packed_pre_gated_delta_rule_backward_match(self):
-        reference_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=False, deterministic_mode=True, conv_kernel_dim=4
-        )
-        fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True, deterministic_mode=False, conv_kernel_dim=4
-        )
-        fused_gdn.load_state_dict(reference_gdn.state_dict())
-
-        batch = 1
-        cu_seqlens = torch.tensor(
-            [0, 1, 4, 6, 11], device=torch.cuda.current_device(), dtype=torch.int32
-        )
-        seq_len = cu_seqlens[-1].item()
-        qkvzba = torch.randn(
-            (seq_len, batch, reference_gdn.in_proj_dim),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        qkvzba_unfused = qkvzba.detach().clone().requires_grad_(True)
-        qkvzba_fused = qkvzba.detach().clone().requires_grad_(True)
-
-        reference_gdn.zero_grad(set_to_none=True)
-        fused_gdn.zero_grad(set_to_none=True)
-
-        unfused_outputs = self._packed_pre_gated_delta_rule_reference(
-            reference_gdn, qkvzba_unfused, cu_seqlens
-        )
-        fused_outputs = fused_gdn._fused_streamed_pre_gated_delta_rule(
-            qkvzba_fused, cu_seqlens_q=cu_seqlens
-        )
-        grad_outputs = [torch.randn_like(output.float()) for output in unfused_outputs]
-
-        unfused_loss = sum(
-            (output.float() * grad).sum() for output, grad in zip(unfused_outputs, grad_outputs)
-        )
-        fused_loss = sum(
-            (output.float() * grad).sum() for output, grad in zip(fused_outputs, grad_outputs)
-        )
-        unfused_loss.backward()
-        fused_loss.backward()
-
-        torch.testing.assert_close(qkvzba_fused.grad, qkvzba_unfused.grad, atol=3e-2, rtol=3e-2)
-        torch.testing.assert_close(
-            fused_gdn.conv1d.weight.grad, reference_gdn.conv1d.weight.grad, atol=3e-2, rtol=3e-2
-        )
-        torch.testing.assert_close(
-            fused_gdn.A_log.grad, reference_gdn.A_log.grad, atol=3e-2, rtol=3e-2
-        )
-        torch.testing.assert_close(
-            fused_gdn.dt_bias.grad, reference_gdn.dt_bias.grad, atol=3e-2, rtol=3e-2
-        )
-
-    def test_fused_and_unfused_packed_pre_gated_delta_rule_backward_repeat4_match(self):
-        reference_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=False,
-            deterministic_mode=True,
-            conv_kernel_dim=4,
-            hidden_size=256,
-            key_head_dim=32,
-            value_head_dim=32,
-            num_key_heads=2,
-            num_value_heads=8,
-        )
-        fused_gdn = self._build_gdn(
-            gdn_pre_gated_delta_rule_fusion=True,
-            deterministic_mode=False,
-            conv_kernel_dim=4,
-            hidden_size=256,
-            key_head_dim=32,
-            value_head_dim=32,
-            num_key_heads=2,
-            num_value_heads=8,
-        )
-        fused_gdn.load_state_dict(reference_gdn.state_dict())
-
-        batch = 1
-        cu_seqlens = torch.tensor(
-            [0, 3, 8, 16], device=torch.cuda.current_device(), dtype=torch.int32
-        )
-        seq_len = cu_seqlens[-1].item()
-        qkvzba = torch.randn(
-            (seq_len, batch, reference_gdn.in_proj_dim),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        qkvzba_unfused = qkvzba.detach().clone().requires_grad_(True)
-        qkvzba_fused = qkvzba.detach().clone().requires_grad_(True)
-
-        reference_gdn.zero_grad(set_to_none=True)
-        fused_gdn.zero_grad(set_to_none=True)
-
-        unfused_outputs = self._packed_pre_gated_delta_rule_reference(
-            reference_gdn, qkvzba_unfused, cu_seqlens
-        )
-        fused_outputs = fused_gdn._fused_streamed_pre_gated_delta_rule(
-            qkvzba_fused, cu_seqlens_q=cu_seqlens
-        )
-        grad_outputs = [torch.randn_like(output.float()) for output in unfused_outputs]
-
-        unfused_loss = sum(
-            (output.float() * grad).sum() for output, grad in zip(unfused_outputs, grad_outputs)
-        )
-        fused_loss = sum(
-            (output.float() * grad).sum() for output, grad in zip(fused_outputs, grad_outputs)
-        )
-        unfused_loss.backward()
-        fused_loss.backward()
-
-        torch.testing.assert_close(qkvzba_fused.grad, qkvzba_unfused.grad, atol=3e-2, rtol=3e-2)
-        torch.testing.assert_close(
-            fused_gdn.conv1d.weight.grad, reference_gdn.conv1d.weight.grad, atol=3e-2, rtol=3e-2
-        )
-        torch.testing.assert_close(
-            fused_gdn.A_log.grad, reference_gdn.A_log.grad, atol=3e-2, rtol=3e-2
-        )
-        torch.testing.assert_close(
-            fused_gdn.dt_bias.grad, reference_gdn.dt_bias.grad, atol=3e-2, rtol=3e-2
-        )
-
-    def test_fused_packed_conv_forward_boundary_isolation(self):
-        from megatron.core.fusions.fused_pre_gated_delta_rule import (
-            fused_streamed_pre_gated_delta_rule,
-        )
-
-        seq_len = 5
-        boundary = 3
-        num_key_heads = 1
-        num_value_heads = 4
-        key_head_dim = 32
-        value_head_dim = 32
-        conv_width = 4
-        qk_channels = num_key_heads * key_head_dim
-        v_channels = num_value_heads * value_head_dim
-        k_offset = qk_channels
-        v_offset = 2 * qk_channels
-        total_channels = 2 * qk_channels + 2 * v_channels + 2 * num_value_heads
-        device = torch.cuda.current_device()
-
-        qkvzba = torch.zeros((seq_len, 1, total_channels), device=device, dtype=torch.bfloat16)
-        qkvzba[boundary - 1, 0, :qk_channels] = 10.0
-        qkvzba[boundary - 1, 0, k_offset : k_offset + qk_channels] = 10.0
-        qkvzba[boundary - 1, 0, v_offset : v_offset + v_channels] = 10.0
-        conv_weight = torch.zeros((2 * qk_channels + v_channels, 1, conv_width), device=device)
-        conv_weight[:qk_channels, 0, conv_width - 2] = 1.0
-        conv_weight[k_offset : k_offset + qk_channels, 0, conv_width - 2] = 1.0
-        conv_weight[v_offset : v_offset + v_channels, 0, conv_width - 2] = 1.0
-        A_log = torch.zeros((num_value_heads,), device=device, dtype=torch.bfloat16)
-        dt_bias = torch.zeros((num_value_heads,), device=device, dtype=torch.bfloat16)
-        cu_seqlens = torch.tensor([0, boundary, seq_len], device=device, dtype=torch.int32)
-
-        query, key, value, _, _, _ = fused_streamed_pre_gated_delta_rule(
-            qkvzba,
-            conv_weight.to(torch.bfloat16),
-            None,
-            A_log,
-            dt_bias,
-            num_key_heads=num_key_heads,
-            num_value_heads=num_value_heads,
-            key_head_dim=key_head_dim,
-            value_head_dim=value_head_dim,
-            cu_seqlens=cu_seqlens,
-        )
-
-        torch.testing.assert_close(
-            query[0, boundary], torch.zeros_like(query[0, boundary]), atol=0.0, rtol=0.0
-        )
-        torch.testing.assert_close(
-            key[0, boundary], torch.zeros_like(key[0, boundary]), atol=0.0, rtol=0.0
-        )
-        torch.testing.assert_close(
-            value[0, boundary], torch.zeros_like(value[0, boundary]), atol=0.0, rtol=0.0
-        )
-
-    def test_fused_packed_conv_backward_boundary_isolation(self):
-        from megatron.core.fusions.fused_pre_gated_delta_rule import (
-            fused_streamed_pre_gated_delta_rule,
-        )
-
-        seq_len = 5
-        boundary = 3
-        num_key_heads = 1
-        num_value_heads = 4
-        key_head_dim = 32
-        value_head_dim = 32
-        conv_width = 4
-        qk_channels = num_key_heads * key_head_dim
-        v_channels = num_value_heads * value_head_dim
-        k_offset = qk_channels
-        v_offset = 2 * qk_channels
-        total_channels = 2 * qk_channels + 2 * v_channels + 2 * num_value_heads
-        device = torch.cuda.current_device()
-
-        qkvzba = torch.zeros(
-            (seq_len, 1, total_channels), device=device, dtype=torch.bfloat16, requires_grad=True
-        )
-        conv_weight = torch.zeros(
-            (2 * qk_channels + v_channels, 1, conv_width),
-            device=device,
-            dtype=torch.bfloat16,
-            requires_grad=True,
-        )
-        with torch.no_grad():
-            qkvzba[boundary - 1, 0, :qk_channels] = 10.0
-            qkvzba[boundary - 1, 0, k_offset : k_offset + qk_channels] = 10.0
-            qkvzba[boundary - 1, 0, v_offset : v_offset + v_channels] = 10.0
-            conv_weight[:qk_channels, 0, conv_width - 2] = 1.0
-            conv_weight[k_offset : k_offset + qk_channels, 0, conv_width - 2] = 1.0
-            conv_weight[v_offset : v_offset + v_channels, 0, conv_width - 2] = 1.0
-        A_log = torch.zeros(
-            (num_value_heads,), device=device, dtype=torch.bfloat16, requires_grad=True
-        )
-        dt_bias = torch.zeros(
-            (num_value_heads,), device=device, dtype=torch.bfloat16, requires_grad=True
-        )
-        cu_seqlens = torch.tensor([0, boundary, seq_len], device=device, dtype=torch.int32)
-
-        query, key, value, gate, beta, g = fused_streamed_pre_gated_delta_rule(
-            qkvzba,
-            conv_weight,
-            None,
-            A_log,
-            dt_bias,
-            num_key_heads=num_key_heads,
-            num_value_heads=num_value_heads,
-            key_head_dim=key_head_dim,
-            value_head_dim=value_head_dim,
-            cu_seqlens=cu_seqlens,
-        )
-
-        loss = (
-            query[0, boundary].float().sum()
-            + key[0, boundary].float().sum()
-            + value[0, boundary].float().sum()
-        )
-        loss = loss + 0.0 * (gate.float().sum() + beta.float().sum() + g.float().sum())
-        loss.backward()
-
-        leaked_q_grad = qkvzba.grad[boundary - 1, 0, :qk_channels]
-        leaked_k_grad = qkvzba.grad[boundary - 1, 0, k_offset : k_offset + qk_channels]
-        leaked_v_grad = qkvzba.grad[boundary - 1, 0, v_offset : v_offset + v_channels]
-        torch.testing.assert_close(
-            leaked_q_grad, torch.zeros_like(leaked_q_grad), atol=0.0, rtol=0.0
-        )
-        torch.testing.assert_close(
-            leaked_k_grad, torch.zeros_like(leaked_k_grad), atol=0.0, rtol=0.0
-        )
-        torch.testing.assert_close(
-            leaked_v_grad, torch.zeros_like(leaked_v_grad), atol=0.0, rtol=0.0
-        )
-
-
-@pytest.mark.skipif(not HAVE_FLA, reason="FLA is not installed.")
-@pytest.mark.skipif(not HAVE_FUSED_PRE_GDR, reason="causal-conv1d fused backward is not installed.")
-@pytest.mark.internal
-class TestFusedPreGatedDeltaRuleChunkwiseCP:
-
-    @pytest.fixture(scope='function', autouse=True)
-    def setup_method(self):
-        Utils.initialize_model_parallel(
-            tensor_model_parallel_size=1, pipeline_model_parallel_size=1, context_parallel_size=2
-        )
-        model_parallel_cuda_manual_seed(123)
-
-        tp_group = parallel_state.get_tensor_model_parallel_group()
-        cp_group = parallel_state.get_context_parallel_group()
-        self.pg_collection = ProcessGroupCollection(tp=tp_group, cp=cp_group)
-
-    def teardown_method(self):
-        torch.cuda.synchronize()
-        Utils.destroy_model_parallel()
-
-    def _build_gdn(self, *, gdn_pre_gated_delta_rule_fusion: bool, conv_kernel_dim: int = 4):
-        transformer_config = TransformerConfig(
-            hidden_size=256,
-            linear_conv_kernel_dim=conv_kernel_dim,
-            linear_key_head_dim=64,
-            linear_value_head_dim=64,
-            linear_num_key_heads=4,
-            linear_num_value_heads=8,
-            num_layers=1,
-            normalization="RMSNorm",
-            use_cpu_initialization=True,
-            layernorm_zero_centered_gamma=True,
-            num_attention_heads=8,
-            activation_func=F.silu,
-            bf16=True,
-            tensor_model_parallel_size=1,
-            context_parallel_size=2,
-            experimental_attention_variant="gated_delta_net",
-            linear_attention_freq=[1],
-            linear_cp_mode="chunkwise",
-            transformer_impl="transformer_engine",
-            deterministic_mode=False,
-            gdn_pre_gated_delta_rule_fusion=gdn_pre_gated_delta_rule_fusion,
-        )
-        gdn_submodules = get_experimental_attention_variant_module_spec(
-            config=transformer_config
-        ).submodules
-        gdn = GatedDeltaNet(
-            transformer_config,
-            submodules=gdn_submodules,
-            layer_number=1,
-            bias=False,
-            conv_bias=False,
-            conv_init=1.0,
-            use_qk_l2norm=True,
-            A_init_range=(1, 16),
-            pg_collection=self.pg_collection,
-        )
-        return gdn.cuda().bfloat16()
-
-    def _make_hidden_states(self, gdn, *, seq_len_global: int):
-        cp_size = parallel_state.get_context_parallel_world_size()
-        assert seq_len_global % cp_size == 0
-        seq_len_local = seq_len_global // cp_size
-        cp_rank = parallel_state.get_context_parallel_rank()
-        hidden_states = torch.randn(
-            (seq_len_local, 1, gdn.config.hidden_size),
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        return hidden_states + cp_rank
-
-    @staticmethod
-    def _make_packed_seq_params(cu_seqlens):
-        cp_size = parallel_state.get_context_parallel_world_size()
-        cu = torch.tensor(cu_seqlens, device=torch.cuda.current_device(), dtype=torch.int32)
-        return PackedSeqParams(
-            qkv_format='thd',
-            cu_seqlens_q=cu,
-            cu_seqlens_kv=cu,
-            max_seqlen_q=max(cu_seqlens[i + 1] - cu_seqlens[i] for i in range(len(cu_seqlens) - 1)),
-            max_seqlen_kv=max(
-                cu_seqlens[i + 1] - cu_seqlens[i] for i in range(len(cu_seqlens) - 1)
-            ),
-            total_tokens=cu_seqlens[-1] // cp_size,
-        )
-
-    @staticmethod
-    def _run_forward(gdn, hidden_states, packed_seq_params=None):
-        with torch.no_grad():
-            output, bias = gdn(hidden_states, None, packed_seq_params=packed_seq_params)
-        if torch.distributed.is_initialized():
-            torch.distributed.barrier()
-        return output, bias
-
-    @staticmethod
-    def _run_backward(gdn, hidden_states, grad_output, packed_seq_params=None):
-        gdn.zero_grad(set_to_none=True)
-        hidden_states = hidden_states.detach().clone().requires_grad_(True)
-        output, _ = gdn(hidden_states, None, packed_seq_params=packed_seq_params)
-        loss = (output.float() * grad_output).sum()
-        loss.backward()
-        if torch.distributed.is_initialized():
-            torch.distributed.barrier()
-        param_grads = {
-            name: param.grad.detach().clone()
-            for name, param in gdn.named_parameters()
-            if param.grad is not None
-        }
-        return output.detach(), hidden_states.grad.detach().clone(), param_grads
-
-    def test_fused_and_unfused_forward_chunkwise_cp_match(self):
-        unfused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=False)
-        fused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=True)
-        fused_gdn.load_state_dict(unfused_gdn.state_dict())
-        hidden_states = self._make_hidden_states(unfused_gdn, seq_len_global=64)
-
-        unfused_output, unfused_bias = self._run_forward(unfused_gdn, hidden_states)
-        fused_output, fused_bias = self._run_forward(fused_gdn, hidden_states)
-
-        rank = torch.distributed.get_rank()
-        torch.testing.assert_close(
-            fused_output,
-            unfused_output,
-            atol=3e-3,
-            rtol=3e-3,
-            msg=lambda msg: f"chunkwise CP fused forward mismatch ({rank=}): {msg}",
-        )
-        assert fused_bias == unfused_bias
-
-    @pytest.mark.flaky_in_dev
-    def test_fused_and_unfused_backward_chunkwise_cp_match(self):
-        unfused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=False)
-        fused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=True)
-        fused_gdn.load_state_dict(unfused_gdn.state_dict())
-        hidden_states = self._make_hidden_states(unfused_gdn, seq_len_global=64)
-        grad_output = torch.randn(
-            hidden_states.shape, device=torch.cuda.current_device(), dtype=torch.float32
-        )
-
-        unfused_output, unfused_dinput, unfused_grads = self._run_backward(
-            unfused_gdn, hidden_states, grad_output
-        )
-        fused_output, fused_dinput, fused_grads = self._run_backward(
-            fused_gdn, hidden_states, grad_output
-        )
-
-        rank = torch.distributed.get_rank()
-        torch.testing.assert_close(
-            fused_output,
-            unfused_output,
-            atol=3e-3,
-            rtol=3e-3,
-            msg=lambda msg: f"chunkwise CP fused backward output mismatch ({rank=}): {msg}",
-        )
-        torch.testing.assert_close(
-            fused_dinput,
-            unfused_dinput,
-            atol=5e-2,
-            rtol=5e-2,
-            msg=lambda msg: f"chunkwise CP fused input grad mismatch ({rank=}): {msg}",
-        )
-        assert fused_grads.keys() == unfused_grads.keys()
-        for name in unfused_grads:
-            torch.testing.assert_close(
-                fused_grads[name],
-                unfused_grads[name],
-                atol=5e-2,
-                rtol=5e-2,
-                msg=lambda msg, param_name=name: (
-                    f"chunkwise CP fused grad mismatch for {param_name!r} ({rank=}): {msg}"
-                ),
-            )
-
-    def test_fused_chunkwise_cp_rejects_short_local_chunks(self):
-        gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=True, conv_kernel_dim=4)
-        hidden_states = self._make_hidden_states(gdn, seq_len_global=4)
-
-        with pytest.raises((AssertionError, ValueError), match="local.*chunk|conv_kernel_dim"):
-            gdn(hidden_states, None)
-
-    def test_fused_and_unfused_packed_forward_chunkwise_cp_match(self):
-        unfused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=False)
-        fused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=True)
-        fused_gdn.load_state_dict(unfused_gdn.state_dict())
-        packed_seq_params = self._make_packed_seq_params([0, 96, 128])
-        hidden_states = self._make_hidden_states(unfused_gdn, seq_len_global=128)
-
-        unfused_output, unfused_bias = self._run_forward(
-            unfused_gdn, hidden_states, packed_seq_params=packed_seq_params
-        )
-        fused_output, fused_bias = self._run_forward(
-            fused_gdn, hidden_states, packed_seq_params=packed_seq_params
-        )
-
-        rank = torch.distributed.get_rank()
-        torch.testing.assert_close(
-            fused_output,
-            unfused_output,
-            atol=3e-3,
-            rtol=3e-3,
-            msg=lambda msg: f"packed chunkwise CP fused forward mismatch ({rank=}): {msg}",
-        )
-        assert fused_bias == unfused_bias
-
-    @pytest.mark.flaky_in_dev
-    def test_fused_and_unfused_packed_backward_chunkwise_cp_match(self):
-        unfused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=False)
-        fused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=True)
-        fused_gdn.load_state_dict(unfused_gdn.state_dict())
-        packed_seq_params = self._make_packed_seq_params([0, 96, 128])
-        hidden_states = self._make_hidden_states(unfused_gdn, seq_len_global=128)
-        grad_output = torch.randn(
-            hidden_states.shape, device=torch.cuda.current_device(), dtype=torch.float32
-        )
-
-        unfused_output, unfused_dinput, unfused_grads = self._run_backward(
-            unfused_gdn, hidden_states, grad_output, packed_seq_params=packed_seq_params
-        )
-        fused_output, fused_dinput, fused_grads = self._run_backward(
-            fused_gdn, hidden_states, grad_output, packed_seq_params=packed_seq_params
-        )
-
-        rank = torch.distributed.get_rank()
-        torch.testing.assert_close(
-            fused_output,
-            unfused_output,
-            atol=3e-3,
-            rtol=3e-3,
-            msg=lambda msg: f"packed chunkwise CP fused backward output mismatch ({rank=}): {msg}",
-        )
-        torch.testing.assert_close(
-            fused_dinput,
-            unfused_dinput,
-            atol=5e-2,
-            rtol=5e-2,
-            msg=lambda msg: f"packed chunkwise CP fused input grad mismatch ({rank=}): {msg}",
-        )
-        assert fused_grads.keys() == unfused_grads.keys()
-        # Packed chunkwise CP compares two bf16 backward implementations that
-        # differ in their causal-conv boundary path. A few parameter-gradient
-        # elements can land just above the dense-path tolerance while the fused
-        # output and input gradients remain tightly matched.
-        packed_param_grad_atol = 1e-1
-        for name in unfused_grads:
-            torch.testing.assert_close(
-                fused_grads[name],
-                unfused_grads[name],
-                atol=packed_param_grad_atol,
-                rtol=5e-2,
-                msg=lambda msg, param_name=name: (
-                    f"packed chunkwise CP fused grad mismatch for {param_name!r} ({rank=}): {msg}"
-                ),
-            )
-
-    def test_fused_and_unfused_packed_partial_boundary_chunkwise_cp_match(self):
-        unfused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=False)
-        fused_gdn = self._build_gdn(gdn_pre_gated_delta_rule_fusion=True)
-        fused_gdn.load_state_dict(unfused_gdn.state_dict())
-        packed_seq_params = self._make_packed_seq_params([0, 4, 12])
-        hidden_states = self._make_hidden_states(unfused_gdn, seq_len_global=12)
-        grad_output = torch.randn(
-            hidden_states.shape, device=torch.cuda.current_device(), dtype=torch.float32
-        )
-
-        unfused_output, unfused_dinput, unfused_grads = self._run_backward(
-            unfused_gdn, hidden_states, grad_output, packed_seq_params=packed_seq_params
-        )
-        fused_output, fused_dinput, fused_grads = self._run_backward(
-            fused_gdn, hidden_states, grad_output, packed_seq_params=packed_seq_params
-        )
-
-        rank = torch.distributed.get_rank()
-        torch.testing.assert_close(
-            fused_output,
-            unfused_output,
-            atol=3e-3,
-            rtol=3e-3,
-            msg=lambda msg: f"partial-boundary packed CP fused output mismatch ({rank=}): {msg}",
-        )
-        torch.testing.assert_close(
-            fused_dinput,
-            unfused_dinput,
-            atol=5e-2,
-            rtol=5e-2,
-            msg=lambda msg: f"partial-boundary packed CP fused input grad mismatch ({rank=}): {msg}",
-        )
-        assert fused_grads.keys() == unfused_grads.keys()
-        # Packed chunkwise CP compares two bf16 backward implementations that
-        # differ in their causal-conv boundary path. A few parameter-gradient
-        # elements can land just above the dense-path tolerance while the fused
-        # output and input gradients remain tightly matched.
-        packed_param_grad_atol = 1e-1
-        for name in unfused_grads:
-            torch.testing.assert_close(
-                fused_grads[name],
-                unfused_grads[name],
-                atol=packed_param_grad_atol,
-                rtol=5e-2,
-                msg=lambda msg, param_name=name: (
-                    f"partial-boundary packed CP fused grad mismatch for {param_name!r} "
-                    f"({rank=}): {msg}"
-                ),
-            )
 
 
 @pytest.mark.skipif(not HAVE_FLA, reason="FLA is not installed.")
@@ -1669,22 +503,17 @@ class TestGDNCuSeqlensResolve:
 
 @pytest.mark.parametrize("sequence_packing", [False, True])
 @pytest.mark.parametrize(
-    ("tp", "sp", "cp", "linear_cp_mode"),
+    ("tp", "sp", "cp"),
     [
-        (4, False, 1, None),  # TP w/o SP
-        (4, True, 1, None),  # TP w/ SP
-        (1, False, 2, "headwise"),  # Headwise CP
-        (2, False, 2, "headwise"),  # TP w/o SP + Headwise CP
-        (2, True, 2, "headwise"),  # TP w/ SP + Headwise CP
-        (1, False, 2, "chunkwise"),  # Chunkwise CP
-        (2, False, 2, "chunkwise"),  # TP w/o SP + chunkwise CP
-        (2, True, 2, "chunkwise"),  # TP w/ SP + chunkwise CP
+        (4, False, 1),  # TP w/o SP
+        (4, True, 1),  # TP w/ SP
+        (1, False, 2),  # CP
+        (2, False, 2),  # TP w/o SP + CP
+        (2, True, 2),  # TP w/ SP + CP
     ],
 )
 @pytest.mark.skipif(not HAVE_FLA, reason="FLA is not installed.")
-def test_parallel_gated_delta_net_correctness(
-    tmp_path_dist_ckpt, sequence_packing, tp, sp, cp, linear_cp_mode
-):
+def test_parallel_gated_delta_net_correctness(tmp_path_dist_ckpt, sequence_packing, tp, sp, cp):
     transformer_config = TransformerConfig(
         hidden_size=128,
         linear_conv_kernel_dim=2,
@@ -1701,7 +530,6 @@ def test_parallel_gated_delta_net_correctness(
         bf16=True,
         experimental_attention_variant="gated_delta_net",
         linear_attention_freq=[1],
-        linear_cp_mode=linear_cp_mode,
         transformer_impl="transformer_engine",
     )
 
@@ -1709,16 +537,10 @@ def test_parallel_gated_delta_net_correctness(
         config=transformer_config, vp_stage=None, pp_rank=0
     )
 
-    cosine_similarity_threshold = None
-    if cp > 1:
-        atol, rtol = 2e-3, 1e-2
-        cosine_similarity_threshold = 0.9999
+    if cp:
+        atol, rtol = 5e-3, 5e-3
     else:
-        atol, rtol = 2e-4, 2e-3
-        cosine_similarity_threshold = 0.99999
-
-    is_chunkwise_cp = linear_cp_mode == "chunkwise" and cp > 1
-    micro_batch_size = 1 if is_chunkwise_cp and not sequence_packing else 4
+        atol, rtol = 5e-4, 5e-4
 
     _test_parallel_attention_correctness(
         transformer_config=transformer_config,
@@ -1726,19 +548,21 @@ def test_parallel_gated_delta_net_correctness(
         tmp_path_dist_ckpt=tmp_path_dist_ckpt,
         atol=atol,
         rtol=rtol,
-        cosine_similarity_threshold=cosine_similarity_threshold,
         tp=tp,
         sp=sp,
         cp=cp,
         seed=123,
         sequence_length=256,
-        micro_batch_size=micro_batch_size,
+        micro_batch_size=4,
         sequence_packing=sequence_packing,
     )
 
 
 @pytest.mark.parametrize("cp_size", [2, 4], scope="class")
 @pytest.mark.internal
+@pytest.mark.skip(
+    "Used to verify the correctness of the fused THD AllToAll implementation, locally validated thus no need to run on CI."
+)
 class TestFusedThdAllToAll:
     """Verify fused 1 AllToAll + permute matches the per-sequence, per-channel loop in GDN."""
